Modularizing codes
