--- src/utilities/ExexutionTimeUtils.java
@@ -1,5 +1,0 @@
-package src.utilities;
-
-public class ExexutionTimeUtils {
-
-}
